Removes metadata repository from the app
