+++ app/src/main/java/org/dhis2/AppComponent.java
@@ -1,8 +1,6 @@
 package org.dhis2;
 
 import org.dhis2.data.database.DbModule;
-import org.dhis2.data.metadata.MetadataModule;
-import org.dhis2.data.qr.QRModule;
 import org.dhis2.data.schedulers.SchedulerModule;
 import org.dhis2.data.server.ServerComponent;
 import org.dhis2.data.server.ServerModule;
@@ -10,8 +8,6 @@ import org.dhis2.usescases.login.LoginComponent;
 import org.dhis2.usescases.login.LoginModule;
 import org.dhis2.usescases.splash.SplashComponent;
 import org.dhis2.usescases.splash.SplashModule;
-import org.dhis2.usescases.sync.SyncComponent;
-import org.dhis2.usescases.sync.SyncModule;
 import org.dhis2.utils.UtilsModule;
 
 import javax.inject.Singleton;
@@ -23,7 +19,7 @@ import dagger.Component;
  */
 @Singleton
 @Component(modules = {
-        AppModule.class, DbModule.class, SchedulerModule.class, UtilsModule.class, MetadataModule.class
+        AppModule.class, DbModule.class, SchedulerModule.class, UtilsModule.class
 })
 public interface AppComponent {
 
@@ -36,8 +32,6 @@ public interface AppComponent {
         Builder schedulerModule(SchedulerModule schedulerModule);
 
         Builder utilModule(UtilsModule utilsModule);
-
-        Builder metadataModule(MetadataModule metadataModule);
 
         AppComponent build();
         //ter
