+++ app/src/main/java/org/dhis2/AppModule.java
@@ -1,6 +1,7 @@
 package org.dhis2;
 
 import android.content.Context;
+
 import androidx.annotation.NonNull;
 
 import org.apache.commons.jexl2.JexlEngine;
