--- app/src/main/java/org/dhis2/data/metadata/MetadataRepositoryImpl.java
@@ -1,43 +1,0 @@
-package org.dhis2.data.metadata;
-
-import androidx.annotation.NonNull;
-
-import com.squareup.sqlbrite2.BriteDatabase;
-
-import org.hisp.dhis.android.core.event.EventModel;
-import org.hisp.dhis.android.core.program.ProgramModel;
-
-import io.reactivex.Observable;
-
-
-/**
- * QUADRAM. Created by ppajuelo on 04/12/2017.
- */
-
-public class MetadataRepositoryImpl implements MetadataRepository {
-
-
-    private static final String EXPIRY_DATE_PERIOD_QUERY = String.format(
-            "SELECT program.* FROM %s " +
-                    "JOIN %s ON %s.%s = %s.%s " +
-                    "WHERE %s.%s = ? " +
-                    "LIMIT 1",
-            ProgramModel.TABLE,
-            EventModel.TABLE, ProgramModel.TABLE, ProgramModel.Columns.UID, EventModel.TABLE, EventModel.Columns.PROGRAM,
-            EventModel.TABLE, EventModel.Columns.UID);
-
-    private final BriteDatabase briteDatabase;
-
-    MetadataRepositoryImpl(@NonNull BriteDatabase briteDatabase) {
-        this.briteDatabase = briteDatabase;
-    }
-
-
-    @Override
-    public Observable<ProgramModel> getExpiryDateFromEvent(String eventUid) {
-        return briteDatabase
-                .createQuery(ProgramModel.TABLE, EXPIRY_DATE_PERIOD_QUERY, eventUid == null ? "" : eventUid)
-                .mapToOne(ProgramModel::create);
-    }
-
-}
