+++ app/src/main/java/org/dhis2/usescases/teiDashboard/eventDetail/EventDetailModule.java
@@ -5,7 +5,6 @@ import androidx.annotation.NonNull;
 import com.squareup.sqlbrite2.BriteDatabase;
 
 import org.dhis2.data.dagger.PerActivity;
-import org.dhis2.data.metadata.MetadataRepository;
 import org.dhis2.data.user.UserRepository;
 import org.hisp.dhis.android.core.D2;
 
@@ -35,8 +34,8 @@ public class EventDetailModule {
 
     @Provides
     @PerActivity
-    EventDetailContracts.Presenter providePresenter(EventDetailRepository eventDetailRepository, MetadataRepository metadataRepository, DataEntryStore dataEntryStore) {
-        return new EventDetailPresenter(eventDetailRepository, metadataRepository, dataEntryStore);
+    EventDetailContracts.Presenter providePresenter(EventDetailRepository eventDetailRepository, DataEntryStore dataEntryStore) {
+        return new EventDetailPresenter(eventDetailRepository, dataEntryStore);
     }
 
     @Provides
